backend/requirements.txt
fastapi
uvicorn
torch
opencv-python
numpy
python-multipart
matplotlib
pillow
segmentation-models-pytorch